Fixed errors in test files
