+++ src/app/components/book/book.component.spec.ts
@@ -32,13 +32,12 @@ describe('BookComponent', () => {
 
     expect(title.textContent).toEqual(BookMock.title);
     expect(author.textContent).toEqual(BookMock.author);
-
   });
 
   it('should emit event add book to shopping cart', () => {
     const spyOutput = spyOn(component.addBookToCartEvent, 'emit').and.callThrough();
 
-    const button: HTMLElement = fixture.debugElement.nativeElement.querySelector('#btn-add');
+    const button: HTMLElement = fixture.debugElement.nativeElement.querySelector('.btn-add');
     button.click();
 
     expect(spyOutput).toHaveBeenCalledWith(BookMock);
