+++ src/app/screens/auth/book-list/book-list.component.spec.ts
@@ -6,7 +6,6 @@ import { of } from 'rxjs';
 import { BookModule } from 'src/app/components/book/book.module';
 import { BookMock, BooksFiltered, BooksMockResponse } from 'src/app/helpers/mocks/books.mock';
 import { BooksService } from 'src/app/services/books.service';
-import { ModalService } from 'src/app/services/modal.service';
 import { BookDetailsComponent } from '../book-details/book-details.component';
 
 import { BookListComponent } from './book-list.component';
@@ -16,14 +15,13 @@ describe('BookListComponent', () => {
   let fixture: ComponentFixture<BookListComponent>;
   let booksService: BooksService;
   let router: Router;
-  let modalService: ModalService;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [BookListComponent],
       imports: [HttpClientTestingModule, BookModule,
         RouterTestingModule.withRoutes([{ path: 'auth/book/:id', component: BookDetailsComponent }])],
-      providers: [BooksService, ModalService]
+      providers: [BooksService]
     })
       .compileComponents();
   });
@@ -33,7 +31,6 @@ describe('BookListComponent', () => {
     component = fixture.componentInstance;
     booksService = TestBed.get(BooksService);
     router = TestBed.get(Router);
-    modalService = TestBed.inject(ModalService);
     fixture.detectChanges();
   });
 
@@ -76,13 +73,6 @@ describe('BookListComponent', () => {
     component.showBook(firstBook);
 
     expect(spyRouter).toHaveBeenCalledWith(['auth/book/' + firstBook.id]);
-  });
-
-  it('should show modal of shopping cart', () => {
-    const spyModalService = spyOn(modalService, 'open').and.callThrough();
-    component.showShoppingCart();
-
-    expect(spyModalService).toHaveBeenCalled();
   });
 
   it('should add book to list of shopping cart', () => {
